Rename monomers to analyzer. Update jsdoc a bit.
